Fix repeated task resolving in single question tasks, refactor single question tasks conversation.
